# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexSift Publisher
  version: 2.3.0
  description: Publishes validated pt-BR NexSift signals and lists recently published signals.
servers:
  - url: https://oh4qzz4ivd.execute-api.us-east-1.amazonaws.com/
paths:
  /:
    get:
      operationId: listRecentPosts
      summary: List the most recently published signals
      security:
        - bearerAuth: []
      parameters:
        - name: since
          in: query
          required: false
          description: Only signals published at or after this instant (ISO 8601).
          schema:
            type: string
            format: date-time
        - name: topic
          in: query
          required: false
          description: Only signals whose single topic equals this value.
          schema:
            type: string
            enum: [ai, development, cloud, devops, security, industry, design]
        - name: signalType
          in: query
          required: false
          schema:
            type: string
            enum: [release, risk, shift, research, industry, opportunity]
        - name: query
          in: query
          required: false
          description: Case-insensitive substring match over title, description and tags.
          schema:
            type: string
        - name: tag
          in: query
          required: false
          description: Exact tag match, case-insensitive.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 30
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Pagination offset applied after all filters.
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: detail
          in: query
          required: false
          description: Full summaries for UI-like consumers or compact editorial items for coverage/discovery.
          schema:
            type: string
            enum: [full, compact]
            default: full
      responses:
        '200':
          description: Signals from the complete archive with `total` counting all matches before pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  posts:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/PostSummary'
                        - $ref: '#/components/schemas/PostListItem'
        '401':
          description: Invalid publication token
        '503':
          description: Recent-list retrieval failed temporarily
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: publishPost
      summary: Publish or update a NexSift signal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [post]
              properties:
                post:
                  $ref: '#/components/schemas/PostDraft'
      responses:
        '201':
          description: Signal published or updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  slug:
                    type: string
                  operation:
                    type: string
                    enum: [created, updated]
                  publishedAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          description: Malformed JSON payload
        '401':
          description: Invalid publication token
        '409':
          description: A slug-less creation resolved to an existing signal; retry as an explicit update with that slug
        '422':
          description: Invalid payload, editorial gate not met, source verification failed, or cover image rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Retryable upstream/source/image failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '504':
          description: Upstream timeout during verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Publication failed
  /validate-source:
    post:
      operationId: validateSource
      summary: Open and validate a candidate source URL
      description: Follows redirects (blocking private and metadata hosts), inspects the page and records the outcome. A URL is only publishable when the check succeeds.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Source is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceCheck'
        '401':
          description: Invalid publication token
        '422':
          description: Source rejected permanently (404, soft-404, homepage redirect, blocked host or unsuitable content)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Retryable upstream failure while validating the source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '504':
          description: Upstream timeout while validating the source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /audit-sources:
    post:
      operationId: auditSources
      summary: Revalidate every source of every published signal
      description: Periodic routine for link rot detection. Updates the stored verification record of each source without touching editorial content.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Audit completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkedAt:
                    type: string
                    format: date-time
                  checked:
                    type: integer
                  healthy:
                    type: integer
                  redirected:
                    type: integer
                  temporarily_unavailable:
                    type: integer
                  broken:
                    type: integer
                  replaced:
                    type: integer
                  blocked:
                    type: integer
                  replacements:
                    type: integer
                  replacementCandidates:
                    type: integer
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/SourceAuditEntry'
        '401':
          description: Invalid publication token
  /posts/resolve:
    post:
      operationId: resolvePost
      summary: Resolve a signal identity with the backend slug function
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostIdentity'
      responses:
        '200':
          description: Signal identity resolved
          content:
            application/json:
              schema:
                type: object
                required: [exists, slug]
                properties:
                  exists:
                    type: boolean
                  slug:
                    type: string
                  post:
                    $ref: '#/components/schemas/Post'
        '401':
          description: Invalid publication token
        '422':
          description: Invalid identity payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /posts/{slug}:
    get:
      operationId: getPost
      summary: Get the full signal by slug
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: ^(ai|development|cloud|devops|security|industry|design)-[a-z0-9]+(-[a-z0-9]+)*-\d{4}-\d{2}-\d{2}$
      responses:
        '200':
          description: Full signal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          description: Invalid publication token
        '404':
          description: Signal not found
    delete:
      operationId: deletePost
      summary: Delete a signal by slug
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: ^(ai|development|cloud|devops|security|industry|design)-[a-z0-9]+(-[a-z0-9]+)*-\d{4}-\d{2}-\d{2}$
      responses:
        '200':
          description: Signal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  slug:
                    type: string
                  deletedAt:
                    type: string
                    format: date-time
        '401':
          description: Invalid publication token
        '404':
          description: Signal not found
  /posts/{slug}/copy:
    patch:
      operationId: revisePostCopy
      summary: Revise the wording of an existing signal
      description: Only wording fields may change. Identity, sources, images, scores, publishedAt and updatedAt are preserved. Use publishPost for material changes.
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [description, content, whyItMatters, whatToWatch]
              properties:
                description:
                  type: string
                  minLength: 30
                  maxLength: 260
                content:
                  type: string
                  minLength: 100
                whyItMatters:
                  type: string
                  minLength: 30
                  maxLength: 800
                whatToWatch:
                  type: string
                  minLength: 30
                  maxLength: 500
      responses:
        '200':
          description: Copy revised without changing publication or material-update timestamps
        '401':
          description: Invalid publication token
        '404':
          description: Signal not found
        '422':
          description: Invalid wording payload
  /posts/{slug}/sources/{index}/replace:
    post:
      operationId: replaceSource
      summary: Replace one source after mechanically validating the new URL
      description: The new URL is opened and validated by the backend before the replacement is recorded. The original URL stays in the replacement history, so link rot can be distinguished from a source that never existed.
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: ^(ai|development|cloud|devops|security|industry|design)-[a-z0-9]+(-[a-z0-9]+)*-\d{4}-\d{2}-\d{2}$
        - name: index
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newUrl, reason]
              properties:
                newUrl:
                  type: string
                  format: uri
                reason:
                  type: string
      responses:
        '200':
          description: Updated signal with the replaced source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          description: Invalid publication token
        '404':
          description: Signal not found
        '422':
          description: Invalid newUrl, missing reason or source index out of range
        '503':
          description: Retryable upstream failure while validating the replacement source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '504':
          description: Upstream timeout while validating the replacement source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ApiError:
      type: object
      required: [error, code, requestId, correlationId]
      properties:
        error:
          type: string
        code:
          type: string
        requestId:
          type: string
        correlationId:
          type: string
        retryable:
          type: boolean
        reason:
          type: string
        attempt:
          type: integer
        upstreamStatus:
          type: integer
          nullable: true
        check:
          $ref: '#/components/schemas/SourceCheck'
        issues:
          type: array
          items:
            type: object
            additionalProperties: true
    PostIdentity:
      type: object
      required: [title, topic, signalDate]
      properties:
        title:
          type: string
          minLength: 8
          maxLength: 140
        topic:
          type: string
          enum: [ai, development, cloud, devops, security, industry, design]
        signalDate:
          type: string
          format: date
    PostDraft:
      type: object
      anyOf:
        - required: [slug]
        - required: [title, description, content, whyItMatters, whatToWatch, topic, signalDate, signalType, depth, sources, relevanceScore, confidenceScore]
      properties:
        slug:
          type: string
          description: Optional record identity for updates. When present, the record must already exist and omitted fields are preserved; title changes never recalculate this slug. Without slug, the backend creates a signal and derives {topic}-{slugified title, max 40 chars}-{signalDate}.
          pattern: ^(ai|development|cloud|devops|security|industry|design)-[a-z0-9]+(-[a-z0-9]+)*-\d{4}-\d{2}-\d{2}$
        title:
          type: string
          minLength: 8
          maxLength: 140
        description:
          type: string
          minLength: 30
          maxLength: 260
        content:
          type: string
          minLength: 100
          description: Markdown content in pt-BR. Inline images (![alt](url)) are downloaded, validated and stored by the backend at publication time; the stored content references the local copies.
        whyItMatters:
          type: string
          minLength: 30
          maxLength: 800
        whatToWatch:
          type: string
          minLength: 30
          maxLength: 500
          description: "What to watch next: the concrete follow-up (rollout, adoption, patch, migration, ecosystem response, breaking change)."
        topic:
          type: string
          enum: [ai, development, cloud, devops, security, industry, design]
          description: "The signal's single topic: it defines the slug and is the only topic page where the signal appears."
        relatedTopics:
          type: array
          maxItems: 2
          description: "Contextual topics related to the signal. They never define the slug and never place the signal on their own topic page; they are displayed as context on the signal page."
          items:
            type: string
            enum: [ai, development, cloud, devops, security, industry, design]
        signalDate:
          type: string
          format: date
          description: Actual date of the event, not the publication date.
        publishedAt:
          type: string
          format: date-time
          description: Optional editorial publication instant in ISO 8601. Defaults to the current instant on creation, is preserved when omitted on update, and must not be in the future. Persisted in UTC.
        signalType:
          type: string
          enum: [release, risk, shift, research, industry, opportunity]
        depth:
          type: string
          enum: [practical, deep]
        tags:
          type: array
          maxItems: 10
          items:
            type: string
          default: []
        sources:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PostSource'
        relevanceScore:
          type: number
          minimum: 6.5
          maximum: 10
          description: Editorial gate. Below 6.5 the payload is rejected.
        confidenceScore:
          type: number
          minimum: 7
          maximum: 10
          description: Editorial gate. Below 7 the payload is rejected.
        featured:
          type: boolean
        coverImage:
          $ref: '#/components/schemas/CoverImageDraft'
...
    CoverImage:
      type: object
      required:
        - objectKey
        - sourceUrl
        - contentType
        - alt
        - checkedAt
      properties:
        objectKey:
          type: string
          description: Stored copy key under public/images/ in the content bucket.
        sourceUrl:
          type: string
          format: uri
        contentType:
          type: string
        alt:
          type: string
          minLength: 1
          maxLength: 200
        caption:
          type: string
          maxLength: 300
        checkedAt:
          type: string
          format: date-time
...
    SourceCheck:
      type: object
      properties:
        ok:
          type: boolean
        requestedUrl:
          type: string
        finalUrl:
          type: string
        status:
          type: integer
          nullable: true
        checkedAt:
          type: string
          format: date-time
        pageTitle:
          type: string
          nullable: true
        contentType:
          type: string
          nullable: true
        sourceStatus:
          type: string
          enum: [healthy, redirected, temporarily_unavailable, broken, replaced, blocked]
        attempts:
          type: integer
        retryable:
          type: boolean
        errorCode:
          type: string
          enum: [RATE_LIMITED, UPSTREAM_TIMEOUT, SOURCE_UNAVAILABLE, SOURCE_REJECTED]
    SourceAuditEntry:
      type: object
      properties:
        slug:
          type: string
        url:
          type: string
        requestedUrl:
          type: string
        finalUrl:
          type: string
        httpStatus:
          type: integer
          nullable: true
        sourceStatus:
          type: string
          enum: [healthy, redirected, temporarily_unavailable, broken, replaced, blocked]
        checkedAt:
          type: string
          format: date-time
        needsReplacement:
          type: boolean
          description: True when the source is broken or blocked and no validated replacement is recorded yet.
    PostSummary:
      type: object
      required:
        - id
        - slug
        - title
        - description
        - topic
        - relatedTopics
        - tags
        - signalDate
        - publishedAt
        - readingTime
        - relevanceScore
        - confidenceScore
        - signalType
        - depth
        - locale
        - sources
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        topic:
          type: string
          description: "The signal's single topic; the only topic page where the signal appears."
        relatedTopics:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        signalDate:
          type: string
          format: date
        signalType:
          type: string
        depth:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/PostSource'
        publishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        readingTime:
          type: integer
        relevanceScore:
          type: number
        confidenceScore:
          type: number
        locale:
          type: string
        featured:
          type: boolean
        coverImage:
          $ref: '#/components/schemas/CoverImage'
    PostListItem:
      type: object
      required:
        - slug
        - title
        - description
        - topic
        - relatedTopics
        - signalDate
        - signalType
        - depth
        - publishedAt
        - relevanceScore
        - confidenceScore
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        topic:
          type: string
          description: "The signal's single topic; the only topic page where the signal appears."
        relatedTopics:
          type: array
          items:
            type: string
        signalDate:
          type: string
          format: date
        signalType:
          type: string
        depth:
          type: string
        publishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        relevanceScore:
          type: number
        confidenceScore:
          type: number